feat: Add delete button to contact list header
This commit adds a delete button to the contact list header in the index.html file. The button is added alongside the existing "Update list" button and is wrapped in a flex container. This change allows users to delete the entire contact list with a single click.

The app.js file is also modified to include a new method, `deletarLista()`, in the `ContactManager` class. This method clears the contact list and updates the UI accordingly when the delete button is clicked.

The style.css file is updated to add a new style rule for the delete button. The color of the delete button is set to rgb(156, 17, 17) to make it visually distinct from other buttons.

Note: The previous changes related to updating the contact list are commented out at the end of the app.js file.

Refs: #123